+++ src/ClientContextInterface.php
@@ -39,6 +39,7 @@ use Predis\Command\Argument\TimeSeries\MGetArguments;
 use Predis\Command\Argument\TimeSeries\MRangeArguments;
 use Predis\Command\Argument\TimeSeries\NRangeArguments;
 use Predis\Command\Argument\TimeSeries\RangeArguments;
+use Predis\Command\Argument\TimeSeries\ReadArguments;
 use Predis\Command\CommandInterface;
 use Predis\Command\Container\ACL;
 use Predis\Command\Container\CLIENT;
@@ -143,6 +144,7 @@ use Predis\Command\Redis\VADD;
  * @method $this ftaggregate(string $index, string $query, ?AggregateArguments $arguments = null)
  * @method $this ftaliasadd(string $alias, string $index)
  * @method $this ftaliasdel(string $alias)
+ * @method $this ftaliaslist(string $index)
  * @method $this ftaliasupdate(string $alias, string $index)
  * @method $this ftalter(string $index, FieldInterface[] $schema, ?AlterArguments $arguments = null)
  * @method $this ftcreate(string $index, FieldInterface[] $schema, ?CreateArguments $arguments = null)
@@ -314,7 +316,9 @@ use Predis\Command\Redis\VADD;
  * @method $this tsnrange(array $keys, $fromTimestamp, $toTimestamp, ?NRangeArguments $arguments = null)
  * @method $this tsnrevrange(array $keys, $fromTimestamp, $toTimestamp, ?NRangeArguments $arguments = null)
  * @method $this tsqueryindex(string ...$filterExpression)
+ * @method $this tsquerylabels(?string $label = null, string ...$filterExpression)
  * @method $this tsrange(string $key, $fromTimestamp, $toTimestamp, ?RangeArguments $arguments = null)
+ * @method $this tsread(string $key, $timestamp, ?ReadArguments $arguments = null)
  * @method $this tsrevrange(string $key, $fromTimestamp, $toTimestamp, ?RangeArguments $arguments = null)
  * @method $this xack(string $key, string $group, string ...$id)
  * @method $this xackdel(string $key, string $group, string $mode, array $ids)