+++ tests/Predis/Cluster/PredisStrategyTest.php
@@ -510,6 +510,9 @@ class PredisStrategyTest extends PredisTestCase
             'XDEL' => 'keys-first',
             'XRANGE' => 'keys-first',
 
+            /* commands operating on time series */
+            'TS.READ' => 'keys-first',
+
             /* commands operating on HyperLogLog */
             'PFADD' => 'keys-first',
             'PFCOUNT' => 'keys-all',